Replace `functional` views with regular views containing partials
The purpose of this pattern is to retain the same rendering performance
advantages while also allowing these pages to have their `script` tag
settings. This will further pave the way to using other functional
components within these views.
